// client/src/app/routes/routes.ts
export enum VIEWS {
  // PUBLIC ROUTES
  HOME = "/",
  HOME_WITH_PREFIX = "/home",
  SIGN_UP = "/sign-up",
  SIGN_IN = "/sign-in",

  // USER ROUTES
  USER_DASHBOARD_LAYOUT = "/user",
  USER_DASHBOARD = "/user/dashboard",
  USER_COLLECTIONS = "/user/collections",
  USER_COLLECTION_ITEMS = "/user/collection-items",
  USER_ALL_USERS = "/user/all-users",
}